requirements.txt
requests
python-dotenv
dataclasses-json
typing-extensions
google-cloud-bigquery>=3.0.0